requirements.txt
mkdocs
mkdocs-material
requests
pyyaml
mkdocs-static-i18n
pymdown-extensions
pillow
cairosvg
mkdocs-awesome-pages-plugin
mkdocs-github-admonitions-plugin
mkdocs-minify-html-plugin